Create signing_method.rs
