+++ validator_client/src/initialized_validators.rs
@@ -6,6 +6,7 @@
 //! The `InitializedValidators` struct in this file serves as the source-of-truth of which
 //! validators are managed by this validator client.
 
+use crate::signing_method::SigningMethod;
 use account_utils::{
     read_password, read_password_from_user,
     validator_definitions::{
@@ -23,10 +24,7 @@ use std::fs::File;
 use std::io::{self, Read};
 use std::path::{Path, PathBuf};
 use std::time::Duration;
-use types::{
-    ChainSpec, Domain, Epoch, Fork, Graffiti, Hash256, Keypair, PublicKey, PublicKeyBytes,
-    Signature, SignedRoot,
-};
+use types::{Graffiti, Keypair, PublicKey, PublicKeyBytes};
 
 use crate::key_cache;
 use crate::key_cache::KeyCache;
@@ -79,57 +77,9 @@ pub enum Error {
     UnableToBuildRemoteSignerClient(ReqwestError),
 }
 
-#[derive(Debug, PartialEq)]
-pub enum SigningError {
-    Todo,
-}
-
 impl From<LockfileError> for Error {
     fn from(error: LockfileError) -> Self {
         Self::Lockfile(error)
-    }
-}
-
-/// A method used by a validator to sign messages.
-///
-/// Presently there is only a single variant, however we expect more variants to arise (e.g.,
-/// remote signing).
-pub enum SigningMethod {
-    /// A validator that is defined by an EIP-2335 keystore on the local filesystem.
-    LocalKeystore {
-        voting_keystore_path: PathBuf,
-        voting_keystore_lockfile: Lockfile,
-        voting_keystore: Keystore,
-        voting_keypair: Keypair,
-    },
-    /// A validator that defers to a HTTP server for signing.
-    RemoteSigner {
-        url: Url,
-        http_client: Client,
-        voting_public_key: PublicKey,
-    },
-}
-
-impl SigningMethod {
-    pub fn get_signature<T: SignedRoot>(
-        &self,
-        domain: Domain,
-        data: &T,
-        epoch: Epoch,
-        fork: &Fork,
-        genesis_validators_root: Hash256,
-        spec: &ChainSpec,
-    ) -> Result<Signature, SigningError> {
-        let domain = spec.get_domain(epoch, domain, fork, genesis_validators_root);
-
-        let signing_root = data.signing_root(domain);
-
-        match self {
-            SigningMethod::LocalKeystore { voting_keypair, .. } => {
-                Ok(voting_keypair.sk.sign(signing_root))
-            }
-            SigningMethod::RemoteSigner { .. } => todo!(),
-        }
     }
 }
